+++ packages/discordx/src/decorators/classes/index.ts
@@ -1,0 +1,10 @@
+export * from "./DApplicationCommand.js";
+export * from "./DApplicationCommandGroup.js";
+export * from "./DApplicationCommandOption.js";
+export * from "./DApplicationCommandOptionChoice.js";
+export * from "./DComponent.js";
+export * from "./DDiscord.js";
+export * from "./DGuard.js";
+export * from "./DOn.js";
+export * from "./DSimpleCommand.js";
+export * from "./DSimpleCommandOption.js";
