+++ packages/discordx/src/decorators/index.ts
@@ -1,31 +1,2 @@
-export * from "./classes/DApplicationCommand.js";
-export * from "./classes/DApplicationCommandGroup.js";
-export * from "./classes/DApplicationCommandOption.js";
-export * from "./classes/DApplicationCommandOptionChoice.js";
-export * from "./classes/DComponent.js";
-export * from "./classes/DDiscord.js";
-export * from "./classes/DGuard.js";
-export * from "./classes/DOn.js";
-export * from "./classes/DSimpleCommand.js";
-export * from "./classes/DSimpleCommandOption.js";
-export * from "./decorators/Bot.js";
-export * from "./decorators/ButtonComponent.js";
-export * from "./decorators/ContextMenu.js";
-export * from "./decorators/Discord.js";
-export * from "./decorators/Guard.js";
-export * from "./decorators/Guild.js";
-export * from "./decorators/On.js";
-export * from "./decorators/Once.js";
-export * from "./decorators/Permission.js";
-export * from "./decorators/SelectMenuComponent.js";
-export * from "./decorators/SimpleCommand.js";
-export * from "./decorators/SimpleCommandOption.js";
-export * from "./decorators/Slash.js";
-export * from "./decorators/SlashChoice.js";
-export * from "./decorators/SlashGroup.js";
-export * from "./decorators/SlashOption.js";
-export * from "./params/djs.js";
-export * from "./params/EventParams.js";
-export * from "./params/SimpleCommandParams.js";
-export * from "./params/SlashOptionParams.js";
-export * from "./params/SlashParams.js";
+export * from "./classes/index.js";
+export * from "./decorators/index.js";
