--- packages/discordx/src/decorators/params/EventParams.ts
@@ -1,3 +1,0 @@
-export interface EventParams {
-  botIds?: string[];
-}
