--- packages/discordx/src/decorators/params/SlashParams.ts
@@ -1,8 +1,0 @@
-import type { IGuild } from "../../index.js";
-
-export interface ApplicationCommandParams {
-  botIds?: string[];
-  defaultPermission?: boolean;
-  description?: string;
-  guilds?: IGuild[];
-}
