+++ packages/discordx/src/types/index.ts
@@ -1,11 +1,2 @@
-export * from "./core/ClientOptions.js";
-export * from "./core/common.js";
-export * from "./core/DiscordEvents.js";
-export * from "./core/literal.js";
-export * from "./core/Next.js";
-export * from "./core/SlashChoicesType.js";
-export * from "./public/ArgsOf.js";
-export * from "./public/ChannelType.js";
-export * from "./public/CommandType.js";
-export * from "./public/GuardFunction.js";
-export * from "./public/interfaces/index.js";
+export * from "./core/index.js";
+export * from "./public/index.js";
