--- packages/discordx/src/types/public/ChannelType.ts
@@ -1,12 +1,0 @@
-export type ChannelTypes =
-  | "GUILD_TEXT"
-  | "DM"
-  | "GUILD_VOICE"
-  | "GROUP_DM"
-  | "GUILD_CATEGORY"
-  | "GUILD_NEWS"
-  | "GUILD_STORE"
-  | "GUILD_NEWS_THREAD"
-  | "GUILD_PUBLIC_THREAD"
-  | "GUILD_PRIVATE_THREAD"
-  | "GUILD_STAGE_VOICE";
