+++ packages/discordx/src/types/public/index.ts
@@ -1,0 +1,5 @@
+export * from "./CommandType.js";
+export * from "./common.js";
+export * from "./djs.js";
+export * from "./simple command.js";
+export * from "./slash.js";
