--- packages/discordx/src/types/public/interfaces/InitCommandConfig.ts
@@ -1,26 +1,0 @@
-export interface InitCommandConfig {
-  disable?: {
-    /**
-     * Disable the add operation, which registers application commands with Discord
-     */
-    add?: boolean;
-
-    /**
-     * Disable the delete operation, which unregisters application commands with Discord
-     */
-    delete?: boolean;
-
-    /**
-     * Disable the update operation, which update application commands with Discord
-     */
-    update?: boolean;
-  };
-
-  /**
-   * Enable logging
-   */
-  log?: boolean;
-  /**
-   * Disable specific actions
-   */
-}
