--- packages/discordx/src/types/public/interfaces/index.ts
@@ -1,2 +1,0 @@
-export * from "./InitCommandConfig.js";
-export * from "./logger.js";
