--- packages/discordx/src/types/public/interfaces/logger.ts
@@ -1,6 +1,0 @@
-export interface ILogger {
-  error(...args: unknown[]): void;
-  info(...args: unknown[]): void;
-  log(...args: unknown[]): void;
-  warn(...args: unknown[]): void;
-}
